requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
factory-boy==3.2.0
Faker==7.0.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
